Fix: Rename app-creater into app-creator
