+++ src/pages/scripts/packages/app-creator/LICENSE
