+++ src/pages/scripts/packages/app-creator/src/service-manager.js
