+++ src/pages/scripts/ydits-web/services/sounds/sounds.js
@@ -8,7 +8,7 @@
  *
  */
 
-import { Service } from "../../../packages/app-creater/service.js";
+import { Service } from "../../../packages/app-creator/service.js";
 
 /**
  * サウンドを扱う。
